pyproject.toml
[tool.poetry]
name = "smartx-rfid"
version = "2.2.0"
description = "SmartX RFID library"
authors = ["Gabriel Henrique Pascon <gh.pascon@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/ghpascon/smartx_rfid"
repository = "https://github.com/ghpascon/smartx_rfid"
documentation = "https://github.com/ghpascon/smartx_rfid#readme"
keywords = ["python", "library", "RFID", "smartx", "packaging"]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
packages = [{include = "smartx_rfid", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pyserial-asyncio = "0.6"
pyserial = "3.5"
pyepc = "0.5.0"
httpx = "0.28.1"
bleak = "^1.1.1"
pydantic = "^2.12.5"
sqlalchemy = "2.0.29"
pymysql = "1.1.1"
psycopg2 = "^2.9.11"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.0"
twine = "^4.0.0"
ruff = "^0.14.11"

[tool.poetry.group.build.dependencies]
build = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py311"
exclude = ["build", "dist", ".venv", "__pycache__"]  # pastas a ignorar

# Por arquivo
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # permite imports não usados em __init__.py para re-exportações

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "asyncio: marks tests as asyncio tests",
]